backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==2.1.1
Pillow==10.4.0
EasyOCR==1.7.1
rapidfuzz==3.9.6
metaphone==0.6
requests==2.32.3
sqlalchemy==2.0.34
orjson==3.10.7
# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Async Job Processing
celery==5.3.4
redis==5.0.1
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
# Monitoring
prometheus-client==0.19.0
# OpenTelemetry
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
# Database
alembic==1.13.0
asyncpg==0.29.0
# Settings
pydantic-settings==2.1.0
# Load Testing
locust==2.17.0